rename initGame function & change field icon
